Collision with platform detected. Jr no longer falls through platform.

+++ C-Client/src/game/entities/Platform.h
@@ -9,8 +9,8 @@
 #include "Entity.h"
 typedef struct {
     Entity *entity;
-    int width;
-    int height;
+    float width;
+    float height;
 } Platform
 ;
